+++ src/Type/CustomerOrder.php
@@ -17,7 +17,7 @@ namespace MarcusJaschen\Collmex\Type;
  * @property $customer_id
  * @property $customer_salutation
  * @property $customer_title
- * @property $customer_forename
+ * @property $customer_firstname
  * @property $customer_lastname
  * @property $customer_firm
  * @property $customer_department
@@ -68,7 +68,7 @@ namespace MarcusJaschen\Collmex\Type;
  * @property $delivery_conditions_additional
  * @property $delivery_salutation
  * @property $delivery_title
- * @property $delivery_forename
+ * @property $delivery_firstname
  * @property $delivery_lastname
  * @property $delivery_firm
  * @property $delivery_department
@@ -238,7 +238,7 @@ class CustomerOrder extends AbstractType implements TypeInterface
         'customer_id' => null,
         'customer_salutation' => null,
         'customer_title' => null,
-        'customer_forename' => null,
+        'customer_firstname' => null,
         // 10
         'customer_lastname' => null,
         'customer_firm' => null,
@@ -295,7 +295,7 @@ class CustomerOrder extends AbstractType implements TypeInterface
         'delivery_salutation' => null,
         'delivery_title' => null,
         // 60
-        'delivery_forename' => null,
+        'delivery_firstname' => null,
         'delivery_lastname' => null,
         'delivery_firm' => null,
         'delivery_department' => null,